server/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0
livekit>=0.18.0
pydantic>=2.8
aiohttp>=3.9
